+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Replace variables to Group names (#313)
 
+### Changes
+
+- Updated E2E dependencies (#308)
+
 ## [2.6.0] - 2025-04-22
 
 ### Added