+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changes
 
 - Updated E2E dependencies (#308)
+- Updated sort behavior for UI manager and table header cell (#314)
 
 ## [2.6.0] - 2025-04-22